src/flight-service/requirements.txt
Flask==2.2.2
psycopg2_binary==2.9.4
requests==2.28.1
python-dotenv>=0.19.2